+++ core/modules-entry.ts
@@ -1,10 +1,12 @@
 
 import { VueStorefrontModule } from './modules'
+import { Cart } from './modules/cart'
 import { Mailchimp } from './modules/mailchimp'
 import { Mailer } from './modules/mailer'
 // import { Example } from './modules/module-template'
 
 export const enabledModules: VueStorefrontModule[] = [
+  Cart,
   Mailchimp,
   Mailer
   // Example
