+++ core/modules/cart/index.ts
@@ -1,0 +1,11 @@
+import { store } from './store'
+import { VueStorefrontModule, VueStorefrontModuleConfig } from '@vue-storefront/core/modules'
+
+export const KEY = 'cart'
+
+const moduleConfig: VueStorefrontModuleConfig = {
+  key: KEY,
+  store
+}
+
+export const Cart = new VueStorefrontModule(moduleConfig)
