+++ core/modules/cart/store/actions.ts
@@ -1,13 +1,13 @@
 import Vue from 'vue'
 import { ActionTree } from 'vuex'
-import * as types from '../../mutation-types'
-import rootStore from '../../'
+import * as types from './mutation-types'
+import rootStore from '@vue-storefront/store'
 import i18n from '@vue-storefront/i18n'
 import { sha1 } from 'object-hash'
-import { currentStoreView } from '../../lib/multistore'
+import { currentStoreView } from '@vue-storefront/store/lib/multistore'
 import omit from 'lodash-es/omit'
-import RootState from '../../types/RootState'
-import CartState from './types/CartState'
+import RootState from '@vue-storefront/store/types/RootState'
+import CartState from '../types/CartState'
 import isString from 'lodash-es/isString'
 import toString from 'lodash-es/toString'
