+++ core/modules/cart/store/mutations.ts
@@ -1,8 +1,8 @@
 import Vue from 'vue'
 import { MutationTree } from 'vuex'
-import rootStore from '../../'
-import * as types from '../../mutation-types'
-import CartState from './types/CartState'
+import rootStore from '@vue-storefront/store'
+import * as types from './mutation-types'
+import CartState from '../types/CartState'
 
 const mutations: MutationTree<CartState> = {
   /**
